refactor websocket files
